requirements.txt
Django==1.6.5
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.1.1
psycopg2==2.5.3
static3==0.5.1
wsgiref==0.1.2
django-haystack==2.2.0
elasticsearch>=1.0.0,<2.0.0
